Fix bug in login functionality
